spiciermodbus2mqtt is now a package
