+++ __main__.py
@@ -1,0 +1,2 @@
+from .modbus2mqtt import main
+main()
