+++ modbus2mqtt/__main__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+from .modbus2mqtt import main
+main()
